Fix password validation issues.
When updating one's profile, the password length validation fails
because it is a nil value.

The solution is to allow nil values and test for password_digest
presence.

The real solution would be to add a password_confirmation field to the
signup form and use the default `has_secure_password` validations.